+++ customize.sh
@@ -10,11 +10,11 @@
 
 # Uncomment and change 'MINAPI' and 'MAXAPI' to the minimum and maximum android version for your mod
 # Uncomment DYNLIB if you want libs installed to vendor for oreo+ and system for anything older
-# Uncomment DEBUG if you want full debug logs (saved to /sdcard)
+# Uncomment PARTITIONS and list additional partitions you will be modifying (other than system and vendor), for example: PARTITIONS="/odm /product /system_ext"
 #MINAPI=21
 #MAXAPI=25
 #DYNLIB=true
-#DEBUG=true
+#PARTITIONS=""
 
 ##########################################################################################
 # Replace list